Rename AutomatonCalculator to Automaton
